requirements.txt
flake8
python-dotenv
praw
django
google-cloud-firestore
google-cloud-language